+++ class/actions_attachments.class.php
@@ -118,8 +118,7 @@ class ActionsAttachments
 
 			return 0;
 		}
-
-		if ((in_array($action, array('presend', 'send', 'attachments_send', 'confirm_attachments_send')) || preg_match('/^presend/', $action)) && method_exists($object, 'fetchObjectLinked'))
+		if ((in_array($action, array('presend', 'send', 'attachments_send', 'confirm_attachments_send', 'confirm_sendmassmail')) || preg_match('/^presend/', $action)) && method_exists($object, 'fetchObjectLinked'))
 		{
 			require_once DOL_DOCUMENT_ROOT.'/core/lib/files.lib.php';
 
@@ -365,7 +364,6 @@ class ActionsAttachments
 			$this->action = $action;
 
 		}
-
 		return 0;
 	}
 
@@ -385,7 +383,7 @@ class ActionsAttachments
 			return 0; // le cas des tickets est particulier il faut faire une évolution pour le gérer à part
 		}
 
-		if (in_array($this->action, array('presend', 'send')) || preg_match('/^presend/', $this->action))
+		if (in_array($this->action, array('presend', 'send', 'confirm_sendmassmail')) || preg_match('/^presend/', $this->action))
 		{
 			if (!empty($this->TFilePathByTitleKey))
 			{